Fix passing null Block to neighbor notifications during SoulCage update (#24)
The vanilla TileEntity class has the protected `blockType`  field but accessing that
field should be safely done through getBlockType() as the field itself is not assigned
unless called by getBlockType() or some other mod does so. This is a subtle issue
with Sponge as it's attempting to provide as much information about what is going
on during various processes of the game, and one of those processes involves getting
the default block state from the passed Block that is performing the notification. To 
fix this, the fix involves simply calling getBlockType() here, instead of accessing the field
directly.

We can reference the actual line in which Sponge is throwing the NPE here:
https://github.com/SpongePowered/SpongeCommon/blob/bc1bc04d1843776030f6a5b69bea94a2d1d4646e/src/main/java/org/spongepowered/common/event/tracking/PhaseTracker.java#L692
with the provided printout of the error being caught:
https://gist.github.com/gabizou/1a7784a4609882ec26e987d843229861